feat: extract Pixiv API service for better architecture separation
- Create dedicated PixivApiService class in src/services/pixiv-api.ts
- Move fetchArtworkMetadata from downloader.ts to the new API service
- Add custom PixivApiError class for better error handling
- Maintain backward compatibility through re-exports in downloader.ts
- Improve separation of concerns between API calls and download logic
- Add comprehensive JSDoc documentation for API service
- Enhance error handling with detailed error information (status, URL)

Breaking Changes: None - all existing imports continue to work
Benefits:
- Better testability of API functionality
- Clearer code organization and responsibilities
- Foundation for future API enhancements
- Improved error handling and debugging
